inject dependencies into all the modules
