+++ lib/dataStructures/index.js
@@ -1,0 +1,20 @@
+/*!
+ * datastructures-js
+ * index (data structures module object)
+ * Copyright(c) 2015 Eyas Ranjous <eyas@eyasranjous.info>
+ * MIT Licensed
+ */
+
+'use strict';
+
+module.exports = {
+    binarySearchTree: require('./binarySearchTree'),
+    directedGraph: require('./directedGraph'),
+    doublyLinkedList: require('./doublyLinkedList'),
+    linkedList: require('./linkedList'),
+    hashtable: require('./hashtable'),
+    queue: require('./queue'),
+    priorityQueue: require('./priorityQueue'),
+    set: require('./set'),
+    stack: require('./stack')
+};
