+++ lib/iterators/hashtableIterator.js
@@ -7,38 +7,39 @@
 
 'use strict';
 
-var iterator = require('./iterator');
+module.exports = function(iteratorFactory) {
 
-module.exports = function(elements) {
+    return function(elements) {
+        
+        var prototype = iteratorFactory(elements), 
+            self = Object.create(prototype);
 
-    var prototype = iterator(elements), 
-        self = Object.create(prototype);
+        self.chainIterator = null;
 
-    self.chainIterator = null;
+        self.hasNext = function() {
+            var length = Object.keys(elements).length;
+            if ((length > 0 && this.index < length) || 
+                (this.chainIterator && this.chainIterator.hasNext())) {
+                return true;
+            }
+            return false;
+        };
 
-    self.hasNext = function() {
-        var length = Object.keys(elements).length;
-        if ((length > 0 && this.index < length) || 
-            (this.chainIterator && this.chainIterator.hasNext())) {
-            return true;
-        }
-        return false;
+        self.next = function() {
+            if (!this.chainIterator && this.hasNext()) {
+                this.chainIterator = iteratorFactory(prototype.next());
+            }
+            if (this.chainIterator && this.chainIterator.hasNext()) {
+                return this.chainIterator.next();
+            }
+            else if (this.hasNext() && !this.chainIterator.hasNext()) {
+                this.chainIterator = null;
+                return this.next();
+            }
+
+            return null;
+        };
+        
+        return self;
     };
-
-    self.next = function() {
-        if (!this.chainIterator && this.hasNext()) {
-            this.chainIterator = iterator(prototype.next());
-        }
-        if (this.chainIterator && this.chainIterator.hasNext()) {
-            return this.chainIterator.next();
-        }
-        else if (this.hasNext() && !this.chainIterator.hasNext()) {
-            this.chainIterator = null;
-            return this.next();
-        }
-
-        return null;
-    };
-    
-    return self;
 };
