+++ lib/iterators/index.js
@@ -1,0 +1,13 @@
+/*!
+ * datastructures-js
+ * index (iterators module object)
+ * Copyright(c) 2015 Eyas Ranjous <eyas@eyasranjous.info>
+ * MIT Licensed
+ */
+
+'use strict';
+
+module.exports = {
+    iterator: require('./iterator'),
+    hashtableIterator: require('./hashtableIterator')
+};
