+++ lib/nodes/binaryTreeNode.js
@@ -7,27 +7,28 @@
 
 'use strict';
 
-var dllNode = require('./doublyLinkedListNode');
+module.exports = function(dllNodeFactory) {
 
-module.exports = function(left, right, value) {
-    
-    var prototype = dllNode(left, right, value),
-        self = Object.create(prototype);
+    return function(left, right, value) {
+        
+        var prototype = dllNodeFactory(left, right, value),
+            self = Object.create(prototype);
 
-    self.setLeft = self.setPrev;
-    self.setRight = self.setNext;
-    self.getLeft = self.getPrev;
-    self.getRight = self.getNext;
+        self.setLeft = self.setPrev;
+        self.setRight = self.setNext;
+        self.getLeft = self.getPrev;
+        self.getRight = self.getNext;
 
-    // export the binaryTreeNode api (reuse the doublyLinkedListNode export)
-    self.export = function() {
-        var exp = prototype.export();
-        return {
-            getRight: exp.getNext,
-            getLeft: exp.getPrev,
-            getValue: exp.getValue
+        // export the binaryTreeNode api (reuse the doublyLinkedListNode export)
+        self.export = function() {
+            var exp = prototype.export();
+            return {
+                getRight: exp.getNext,
+                getLeft: exp.getPrev,
+                getValue: exp.getValue
+            };
         };
-    };
 
-    return self;
+        return self;
+    };
 };
