+++ lib/nodes/hashTablePair.js
@@ -7,32 +7,35 @@
 
 'use strict';
 
-module.exports = function(key_, value_) {
-    
-    var self = {},
-        value = value_,
-        key = key_;
+module.exports = function() {
 
-    self.getKey = function() {
-        return key;
-    };
+    return function(key_, value_) {
+        
+        var self = {},
+            value = value_,
+            key = key_;
 
-    self.setValue = function(v) {
-        value = v;
-    };
-
-    self.getValue = function() {
-        return value;
-    };
-
-    // export the hashTablePair api
-    self.export = function() {
-        var that = this;
-        return {
-            getKey: that.getKey,
-            getValue: that.getValue
+        self.getKey = function() {
+            return key;
         };
-    };
 
-    return self;
+        self.setValue = function(v) {
+            value = v;
+        };
+
+        self.getValue = function() {
+            return value;
+        };
+
+        // export the hashTablePair api
+        self.export = function() {
+            var that = this;
+            return {
+                getKey: that.getKey,
+                getValue: that.getValue
+            };
+        };
+
+        return self;
+    };
 };
