+++ lib/nodes/index.js
@@ -1,0 +1,15 @@
+/*!
+ * datastructures-js
+ * index (nodes module object)
+ * Copyright(c) 2015 Eyas Ranjous <eyas@eyasranjous.info>
+ * MIT Licensed
+ */
+
+'use strict';
+
+module.exports = {
+    binaryTreeNode: require('./binaryTreeNode'),
+    hashTablePair: require('./hashTablePair'),
+    linkedListNode: require('./linkedListNode'),
+    doublyLinkedListNode: require('./doublyLinkedListNode')
+};
