+++ lib/nodes/linkedListNode.js
@@ -7,49 +7,53 @@
 
 'use strict';
 
-module.exports = function(next_, value_) {
+module.exports = function() {
 
-    var self = {},
-        next = next_,
-        value = value_;
+    return function(next_, value_) {
+        
+        var self = {},
+            next = next_,
+            value = value_;
 
-    self.setNext = function(n) {
-        next = n;
-    };
-    
-    self.setValue = function(v) {
-        value = v;
-    };
-
-    self.getNext = function() {
-        return next;
-    };
-    
-    self.getValue = function() {
-        return value;
-    };
-
-    // export the connected (next) node api
-    self.exportConnectedNode = function(node) {
-        return function() {
-            if (node !== null) {
-                return {
-                    getNext: node.export().getNext,
-                    getValue: node.export().getValue
-                };
-            }
-            return null;
+        self.setNext = function(n) {
+            next = n;
         };
-    };
-
-    // export the linkedListNode api
-    self.export = function() {
-        var that = this; // because context will change in the return
-        return {
-            getNext: that.exportConnectedNode(that.getNext()),
-            getValue: that.getValue
+        
+        self.setValue = function(v) {
+            value = v;
         };
+
+        self.getNext = function() {
+            return next;
+        };
+        
+        self.getValue = function() {
+            return value;
+        };
+
+        // export the connected (next) node api
+        self.exportConnectedNode = function(node) {
+            return function() {
+                if (node !== null) {
+                    return {
+                        getNext: node.export().getNext,
+                        getValue: node.export().getValue
+                    };
+                }
+                return null;
+            };
+        };
+
+        // export the linkedListNode api
+        self.export = function() {
+            var that = this; // because context will change in the return
+            return {
+                getNext: that.exportConnectedNode(that.getNext()),
+                getValue: that.getValue
+            };
+        };
+
+        return self;
     };
 
-    return self;
 };
